feat: add React.lazy for components
