--- src/Components/Main/Pages/MessagesPage/Dialogs/DialogsContainer.js
@@ -1,11 +1,0 @@
-import { Dialogs } from './Dialogs';
-import { useSelector } from 'react-redux';
-import { dialogsSelector } from '../../../../../redux/selectors/messages';
-
-export const DialogsContainer = () => {
-  const dialogs = useSelector(dialogsSelector);
-
-  return (
-    <Dialogs dialogs={dialogs} />
-  );
-};
