+++ src/Components/Main/Pages/ProfilePage/Posts/PostsContainer.js
@@ -1,8 +1,7 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { addPost } from '../../../../../redux/actions/posts';
 import { Posts } from './Posts';
 import { useDispatch, useSelector } from 'react-redux';
-import { useCallback } from 'react';
 import { postsSelector } from '../../../../../redux/selectors/posts';
 import { isAuthSelector } from '../../../../../redux/selectors/auth';
